T4AMatrixCI.jl

Matrix cross interpolation and related factorizations for tensor4all.

API

T4AMatrixCI.MatrixACA — Method

MatrixACA(lu::rrLU{T}) where {T}

Convert a rank-revealing LU decomposition lu into a MatrixACA representation.

T4AMatrixCI.MatrixCI — Type

mutable struct MatrixCrossInterpolation{T}

Represents a cross interpolation of a matrix - or, to be more precise, the data necessary to evaluate a cross interpolation. This data can be fed into various methods such as evaluate(c, i, j) to get interpolated values, and be improved by dynamically adding pivots.

Base.:\ — Method

Override solving Ax = b using LU decomposition

T4AMatrixCI.AinvtimesB — Method

AtimesBinv(A::Matrix, B::Matrix)

Calculates the matrix product $A^{-1} B$, given a square matrix $A$ and a rectangular matrix $B$ in a numerically stable way using QR decomposition. This is useful in case $A^{-1}$ is ill-conditioned.

T4AMatrixCI.AtimesBinv — Method

AtimesBinv(A::Matrix, B::Matrix)

Calculates the matrix product $A B^{-1}$, given a rectangular matrix $A$ and a square matrix $B$ in a numerically stable way using QR decomposition. This is useful in case $B^{-1}$ is ill-conditioned.

T4AMatrixCI.addpivot! — Method

function addpivot!(a::AbstractMatrix{T}, aca::MatrixACA{T})

Find and add a new pivot according to the ACA algorithm in Kumar 2016 ()

T4AMatrixCI.addpivot! — Method

function addpivot(
    a::AbstractMatrix{T},
    ci::MatrixCrossInterpolation{T},
    pivotindices::AbstractArray{T, 2})

Add a new pivot given by pivotindices to the cross interpolation ci of the matrix a.

T4AMatrixCI.addpivot! — Method

function addpivot(
    a::AbstractMatrix{T},
    ci::MatrixCrossInterpolation{T})

Add a new pivot that maximizes the error to the cross interpolation ci of the matrix a.

T4AMatrixCI.crossinterpolate — Method

function crossinterpolate(
    a::AbstractMatrix{T};
    tolerance=1e-6,
    maxiter=200,
    firstpivot=argmax(abs.(a))
) where {T}

Cross interpolate a $m\times n$ matrix $a$, i.e. find a set of indices $I$ and $J$, such that $a(1\ldots m, 1\ldots n) \approx a(1\ldots m, I) (a(I, J))^{-1} a(J, 1\ldots m)$.

T4AMatrixCI.findnewpivot — Method

findnewpivot(
    a::AbstractMatrix{T},
    ci::MatrixCrossInterpolation{T},
    row_indices::Union{Vector{Int}},
    col_indices::Union{Vector{Int}})

Finds the pivot that maximizes the local error across all components of a and ci within the subset given by rowindices and colindices. By default, all avalable rows of ci will be considered.

T4AMatrixCI.lastpivoterror — Method

Correct estimate of the last pivot error A special care is taken for a full-rank matrix: the last pivot error is set to zero.

T4AMatrixCI.localerror — Method

localerror(
    a::AbstractMatrix{T},
    ci::MatrixCrossInterpolation{T},
    row_indices::Union{AbstractVector{Int},Colon,Int}=Colon(),
    col_indices::Union{AbstractVector{Int},Colon,Int}=Colon())

Returns all local errors of the cross interpolation ci with respect to matrix a. To find local errors on a submatrix, specify rowindices or colindices.

T4AMatrixCI.rrlu! — Method

function rrlu!(
    A::AbstractMatrix{T};
    maxrank::Int=typemax(Int),
    reltol::Number=1e-14,
    abstol::Number=0.0,
    leftorthogonal::Bool=true
)::rrLU{T} where {T}

Rank-revealing LU decomposition.

T4AMatrixCI.rrlu — Method

function rrlu(
    A::AbstractMatrix{T};
    maxrank::Int=typemax(Int),
    reltol::Number=1e-14,
    abstol::Number=0.0,
    leftorthogonal::Bool=true
)::rrLU{T} where {T}

Rank-revealing LU decomposition.

T4AMatrixCI.solve — Method

Solve (LU) x = b

L: lower triangular matrix U: upper triangular matrix b: right-hand side vector

Return x

Note: Not optimized for performance

T4AMatrixCI.uk — Method

Compute u_k(x) for all x

T4AMatrixCI.vk — Method

Compute v_k(y) for all y